(feat) move radar_relay to new folder
