+++ hummingbot/client/hummingbot_application.py
@@ -19,7 +19,7 @@ from hummingbot.market.huobi.huobi_market import HuobiMarket
 from hummingbot.market.liquid.liquid_market import LiquidMarket
 from hummingbot.market.eterbase.eterbase_market import EterbaseMarket
 from hummingbot.market.paper_trade import create_paper_trade_market
-from hummingbot.market.radar_relay.radar_relay_market import RadarRelayMarket
+from hummingbot.connector.exchange.radar_relay.radar_relay_market import RadarRelayMarket
 from hummingbot.connector.exchange.bamboo_relay.bamboo_relay_market import BambooRelayMarket
 from hummingbot.market.dolomite.dolomite_market import DolomiteMarket
 from hummingbot.market.kraken.kraken_market import KrakenMarket
