+++ test/integration/test_radar_relay_market.py
@@ -44,7 +44,7 @@ from hummingbot.core.utils.async_utils import (
 from hummingbot.logger import NETWORK
 from hummingbot.market.market_base import OrderType
 from hummingbot.connector.markets_recorder import MarketsRecorder
-from hummingbot.market.radar_relay.radar_relay_market import RadarRelayMarket
+from hummingbot.connector.exchange.radar_relay.radar_relay_market import RadarRelayMarket
 from hummingbot.model.market_state import MarketState
 from hummingbot.model.order import Order
 from hummingbot.model.sql_connection_manager import (
